backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
psycopg2-binary
python-dotenv
werkzeug
web3
eth-account
